refactor: move renderJSONCell to utils file
